Added faker dependency to help with seed data and create more modular seed data files.
