+++ package.json
@@ -22,7 +22,6 @@
     "body-parser": "^1.12.0",
     "bootstrap": "^3.3.4",
     "chai": "^2.1.0",
-    "chai-things": "^0.2.0",
     "chalk": "^1.0.0",
     "connect-session-sequelize": "^3.0.0",
     "cookie-parser": "^1.3.4",
@@ -68,6 +67,7 @@
   },
   "devDependencies": {
     "chai-things": "^0.2.0",
+    "faker": "^3.1.0",
     "karma-phantomjs-launcher": "^1.0.0",
     "ngkookies": "^0.2.0",
     "phantomjs-prebuilt": "^2.1.7",
